requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0